fix missing insurant

+++ src/main/java/de/thkoeln/inf/gpm/vgb/delegate/customer/CalculateBmiDelegate.java
@@ -14,8 +14,8 @@ public class CalculateBmiDelegate implements JavaDelegate {
         Map<String, Object> processVariables;
         processVariables = delegateExecution.getVariables();
 
-        val insurantId = (int) processVariables.get("insurantId");
-        val insurant = Insurant.findById(insurantId);
+        val insurantId = (long) processVariables.get("insurantId");
+        val insurant = Insurant.findById((int) insurantId);
 
         double bmi = calcBmi(insurant.getWeight(), insurant.getSize());
